fix: based on Riyan reviews

+++ Dash/include/Defines/PinDefs.h
@@ -27,7 +27,7 @@ static constexpr int8_t LCD_B5    =  18;
 static constexpr int8_t LCD_B6    =  17;
 static constexpr int8_t LCD_B7    =  10;
 // CAN bus pins
-static constexpr int LCD_CAN_TX = 15;
-static constexpr int LCD_CAN_RX = 16;  
+static constexpr int8_t LCD_CAN_TX = 15;
+static constexpr int8_t LCD_CAN_RX = 16;  
 
 #endif // PIN_DEFS_H
